feat: 사용자 북마크, 영업시간, 편의시설을 포함한 새로운 DTO를 적용하여 매장 요약 및 상세 조회 기능을 구현하고, 엔티티 ID를 int64로 리팩터링.

+++ certifications/entities/certification_photo.go
@@ -1,8 +1,8 @@
 package entities
 
 type CertificationPhoto struct {
-	ID              uint    `gorm:"primaryKey;autoIncrement" json:"id"`
-	CertificationID *uint   `gorm:"column:certificationId" json:"certificationId"`
+	ID              int64   `gorm:"primaryKey;autoIncrement" json:"id"`
+	CertificationID *int64  `gorm:"column:certificationId" json:"certificationId"`
 	ImageName       *string `gorm:"column:imageName;type:varchar(255)" json:"imageName"`
 	SortOrder       int     `gorm:"column:sortOrder;default:0" json:"sortOrder"`
