+++ moderation/entities/moderation_action.go
@@ -7,9 +7,9 @@ import (
 )
 
 type ModerationAction struct {
-	ID        uint      `gorm:"primaryKey;autoIncrement" json:"id"`
-	ReportID  *uint     `gorm:"column:reportId" json:"reportId"`
-	AdminID   *uint     `gorm:"column:adminId" json:"adminId"`
+	ID        int64     `gorm:"primaryKey;autoIncrement" json:"id"`
+	ReportID  *int64    `gorm:"column:reportId" json:"reportId"`
+	AdminID   *int64    `gorm:"column:adminId" json:"adminId"`
 	Action    *string   `gorm:"type:varchar(50)" json:"action"` // ban_user, delete_content, etc
 	Note      *string   `gorm:"type:text" json:"note"`
 	CreatedAt time.Time `gorm:"column:createdAt;type:datetime(6);autoCreateTime" json:"createdAt"`
