+++ notifications/entities/push_subscription.go
@@ -7,8 +7,8 @@ import (
 )
 
 type PushSubscription struct {
-	ID        uint      `gorm:"primaryKey;autoIncrement" json:"id"`
-	UserID    *uint     `gorm:"column:userId" json:"userId"`
+	ID        int64     `gorm:"primaryKey;autoIncrement" json:"id"`
+	UserID    *int64    `gorm:"column:userId" json:"userId"`
 	Platform  string    `gorm:"type:enum('web','android','ios')" json:"platform"`
 	Endpoint  string    `gorm:"type:varchar(500);not null" json:"endpoint"`
 	AuthKey   *string   `gorm:"column:authKey;type:varchar(255)" json:"authKey"`
